:truck: reorganize and fix type issues
