+++ apps/web/src/iframe/spaces.ts
@@ -1,4 +1,4 @@
-import { useLoggedInContext } from "@/data"
+import { useLoggedInContext } from "@/logged-in-context"
 import { useObservable } from "@coat-rack/core/async"
 import { useChannelSubscription } from "@coat-rack/core/messaging"
 import {
