+++ apps/web/src/routes/index.lazy.tsx
@@ -1,4 +1,4 @@
-import { useLoggedInContext } from "@/data"
+import { useLoggedInContext } from "@/logged-in-context"
 import { useObservable } from "@coat-rack/core/async"
 import { Link, createLazyFileRoute } from "@tanstack/react-router"
